Completed timer/stopwatch behaviour

+++ stopwatch-settings.c
@@ -23,7 +23,7 @@ const char *settings_get_initial_value(obs_data_t *settings)
 }
 
 
-uint32_t settings_get_initial_value_as_int(obs_data_t *settings)
+int64_t settings_get_initial_value_as_int(obs_data_t *settings)
 {
 	const char *time_string = settings_get_initial_value(settings);
 	return time_string_to_millis(time_string);
@@ -36,7 +36,7 @@ const char *settings_get_end_value(obs_data_t *settings)
 }
 
 
-uint32_t settings_get_end_value_as_int(obs_data_t *settings)
+int64_t settings_get_end_value_as_int(obs_data_t *settings)
 {
 	const char *time_string = settings_get_end_value(settings);
 	return time_string_to_millis(time_string);
